requirements.txt
requests
psutil
python-dotenv
fastapi
uvicorn
